+++ src/main/java/net/sf/jabref/logic/cleanup/CleanupPreset.java
@@ -15,19 +15,32 @@
 package net.sf.jabref.logic.cleanup;
 
 import net.sf.jabref.JabRefPreferences;
+import net.sf.jabref.exporter.FieldFormatterCleanups;
 
+import java.util.ArrayList;
 import java.util.EnumSet;
+import java.util.Objects;
 
 public class CleanupPreset {
 
     private final EnumSet<CleanupStep> activeJobs;
+    private final FieldFormatterCleanups formatterCleanups;
 
     public CleanupPreset(EnumSet<CleanupStep> activeJobs) {
-        this.activeJobs = activeJobs;
+        this(activeJobs, new FieldFormatterCleanups(false, new ArrayList<>()));
     }
 
     public CleanupPreset(CleanupStep activeJob) {
         this(EnumSet.of(activeJob));
+    }
+
+    public CleanupPreset(FieldFormatterCleanups formatterCleanups) {
+        this(EnumSet.noneOf(CleanupStep.class), formatterCleanups);
+    }
+
+    public CleanupPreset(EnumSet<CleanupStep> activeJobs, FieldFormatterCleanups formatterCleanups) {
+        this.activeJobs = activeJobs;
+        this.formatterCleanups = Objects.requireNonNull(formatterCleanups);
     }
 
     public static CleanupPreset loadFromPreferences(JabRefPreferences preferences) {
@@ -39,15 +52,6 @@ public class CleanupPreset {
         }
         if (preferences.getBoolean(JabRefPreferences.CLEANUP_DOI)) {
             activeJobs.add(CleanupStep.CLEAN_UP_DOI);
-        }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_MONTH)) {
-            activeJobs.add(CleanupStep.CLEAN_UP_MONTH);
-        }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_PAGE_NUMBERS)) {
-            activeJobs.add(CleanupStep.CLEAN_UP_PAGE_NUMBERS);
-        }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_DATE)) {
-            activeJobs.add(CleanupStep.CLEAN_UP_DATE);
         }
         if (preferences.getBoolean(JabRefPreferences.CLEANUP_MAKE_PATHS_RELATIVE)) {
             activeJobs.add(CleanupStep.MAKE_PATHS_RELATIVE);
@@ -61,18 +65,6 @@ public class CleanupPreset {
         if (preferences.getBoolean(JabRefPreferences.CLEANUP_UPGRADE_EXTERNAL_LINKS)) {
             activeJobs.add(CleanupStep.CLEAN_UP_UPGRADE_EXTERNAL_LINKS);
         }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_HTML)) {
-            activeJobs.add(CleanupStep.CONVERT_HTML_TO_LATEX);
-        }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_CASE)) {
-            activeJobs.add(CleanupStep.CONVERT_CASE);
-        }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_LATEX)) {
-            activeJobs.add(CleanupStep.CONVERT_LATEX);
-        }
-        if (preferences.getBoolean(JabRefPreferences.CLEANUP_UNITS)) {
-            activeJobs.add(CleanupStep.CONVERT_UNITS);
-        }
         if (preferences.getBoolean(JabRefPreferences.CLEANUP_UNICODE)) {
             activeJobs.add(CleanupStep.CONVERT_UNICODE_TO_LATEX);
         }
@@ -83,7 +75,10 @@ public class CleanupPreset {
             activeJobs.add(CleanupStep.FIX_FILE_LINKS);
         }
 
-        return new CleanupPreset(activeJobs);
+        FieldFormatterCleanups formatterCleanups = FieldFormatterCleanups.parseFromString(
+                preferences.getStringList(JabRefPreferences.CLEANUP_FORMATTERS));
+
+        return new CleanupPreset(activeJobs, formatterCleanups);
     }
 
     public boolean isCleanUpUpgradeExternalLinks() {
@@ -98,18 +93,6 @@ public class CleanupPreset {
         return isActive(CleanupStep.CLEAN_UP_DOI);
     }
 
-    public boolean isCleanUpMonth() {
-        return isActive(CleanupStep.CLEAN_UP_MONTH);
-    }
-
-    public boolean isCleanUpPageNumbers() {
-        return isActive(CleanupStep.CLEAN_UP_PAGE_NUMBERS);
-    }
-
-    public boolean isCleanUpDate() {
-        return isActive(CleanupStep.CLEAN_UP_DATE);
-    }
-
     public boolean isFixFileLinks() {
         return isActive(CleanupStep.FIX_FILE_LINKS);
     }
@@ -120,22 +103,6 @@ public class CleanupPreset {
 
     public boolean isRenamePDF() {
         return isActive(CleanupStep.RENAME_PDF) || isActive(CleanupStep.RENAME_PDF_ONLY_RELATIVE_PATHS);
-    }
-
-    public boolean isConvertHTMLToLatex() {
-        return isActive(CleanupStep.CONVERT_HTML_TO_LATEX);
-    }
-
-    public boolean isConvertUnits() {
-        return isActive(CleanupStep.CONVERT_UNITS);
-    }
-
-    public boolean isConvertCase() {
-        return isActive(CleanupStep.CONVERT_CASE);
-    }
-
-    public boolean isConvertLaTeX() {
-        return isActive(CleanupStep.CONVERT_LATEX);
     }
 
     public boolean isConvertUnicodeToLatex() {
@@ -154,26 +121,25 @@ public class CleanupPreset {
 
         preferences.putBoolean(JabRefPreferences.CLEANUP_SUPERSCRIPTS, isActive(CleanupStep.CLEAN_UP_SUPERSCRIPTS));
         preferences.putBoolean(JabRefPreferences.CLEANUP_DOI, isActive(CleanupStep.CLEAN_UP_DOI));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_MONTH, isActive(CleanupStep.CLEAN_UP_MONTH));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_PAGE_NUMBERS, isActive(CleanupStep.CLEAN_UP_PAGE_NUMBERS));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_DATE, isActive(CleanupStep.CLEAN_UP_DATE));
         preferences.putBoolean(JabRefPreferences.CLEANUP_MAKE_PATHS_RELATIVE, isActive(CleanupStep.MAKE_PATHS_RELATIVE));
         preferences.putBoolean(JabRefPreferences.CLEANUP_RENAME_PDF, isActive(CleanupStep.RENAME_PDF));
         preferences.putBoolean(JabRefPreferences.CLEANUP_RENAME_PDF_ONLY_RELATIVE_PATHS,
                 isActive(CleanupStep.RENAME_PDF_ONLY_RELATIVE_PATHS));
         preferences.putBoolean(JabRefPreferences.CLEANUP_UPGRADE_EXTERNAL_LINKS,
                 isActive(CleanupStep.CLEAN_UP_UPGRADE_EXTERNAL_LINKS));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_HTML, isActive(CleanupStep.CONVERT_HTML_TO_LATEX));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_CASE, isActive(CleanupStep.CONVERT_CASE));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_LATEX, isActive(CleanupStep.CONVERT_LATEX));
-        preferences.putBoolean(JabRefPreferences.CLEANUP_UNITS, isActive(CleanupStep.CONVERT_UNITS));
         preferences.putBoolean(JabRefPreferences.CLEANUP_UNICODE, isActive(CleanupStep.CONVERT_UNICODE_TO_LATEX));
         preferences.putBoolean(JabRefPreferences.CLEANUP_CONVERT_TO_BIBLATEX, isActive(CleanupStep.CONVERT_TO_BIBLATEX));
         preferences.putBoolean(JabRefPreferences.CLEANUP_FIX_FILE_LINKS, isActive(CleanupStep.FIX_FILE_LINKS));
+
+        preferences.putStringList(JabRefPreferences.CLEANUP_FORMATTERS, formatterCleanups.convertToString());
     }
 
     private Boolean isActive(CleanupStep step) {
         return activeJobs.contains(step);
+    }
+
+    public FieldFormatterCleanups getFormatterCleanups() {
+        return formatterCleanups;
     }
 
     public enum CleanupStep {
@@ -185,12 +151,6 @@ public class CleanupPreset {
          * Removes the http://... for each DOI. Moves DOIs from URL and NOTE filed to DOI field.
          */
         CLEAN_UP_DOI,
-        CLEAN_UP_MONTH,
-        CLEAN_UP_PAGE_NUMBERS,
-        /**
-         * Format dates correctly (yyyy-mm-dd or yyyy-mm)
-         */
-        CLEAN_UP_DATE,
         MAKE_PATHS_RELATIVE,
         RENAME_PDF,
         RENAME_PDF_ONLY_RELATIVE_PATHS,
@@ -198,16 +158,6 @@ public class CleanupPreset {
          * Collects file links from the pdf or ps field, and adds them to the list contained in the file field.
          */
         CLEAN_UP_UPGRADE_EXTERNAL_LINKS,
-        /**
-         * Converts HTML code to LaTeX code
-         */
-        CONVERT_HTML_TO_LATEX,
-        /**
-         * Adds curly brackets {} around keywords
-         */
-        CONVERT_CASE,
-        CONVERT_LATEX,
-        CONVERT_UNITS,
         /**
          * Converts Unicode characters to LaTeX code
          */