+++ src/test/java/net/sf/jabref/exporter/FieldFormatterCleanupsTest.java
@@ -3,6 +3,7 @@ package net.sf.jabref.exporter;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.logic.cleanup.FieldFormatterCleanup;
+import net.sf.jabref.logic.formatter.Formatter;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.BibtexEntryTypes;
 import org.junit.Before;
@@ -11,11 +12,14 @@ import org.junit.Test;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
-public class SaveActionsTest {
+public class FieldFormatterCleanupsTest {
 
     private BibEntry entry;
 
@@ -44,7 +48,7 @@ public class SaveActionsTest {
     @Test
     public void checkSimpleUseCase() throws IOException {
 
-        SaveActions actions = new SaveActions(true, "title[IdentityFormatter]");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "title[IdentityFormatter]");
 
         actions.applySaveActions(entry);
 
@@ -54,7 +58,7 @@ public class SaveActionsTest {
     @Test
     public void invalidSaveActionSting() throws IOException {
 
-        SaveActions actions = new SaveActions(true, "title");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "title");
 
         actions.applySaveActions(entry);
 
@@ -64,7 +68,7 @@ public class SaveActionsTest {
     @Test
     public void checkLowerCaseSaveAction() throws IOException {
 
-        SaveActions actions = new SaveActions(true, "title[LowerCaseChanger]");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "title[LowerCaseChanger]");
 
         actions.applySaveActions(entry);
 
@@ -73,7 +77,7 @@ public class SaveActionsTest {
 
     @Test
     public void checkTwoSaveActionsForOneField() throws IOException {
-        SaveActions actions = new SaveActions(true, "title[LowerCaseChanger,IdentityFormatter]");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "title[LowerCaseChanger,IdentityFormatter]");
 
         assertEquals(2, actions.getConfiguredActions().size());
 
@@ -85,7 +89,7 @@ public class SaveActionsTest {
     @Test
     public void checkThreeSaveActionsForOneField() throws IOException {
 
-        SaveActions actions = new SaveActions(true, "title[LowerCaseChanger,IdentityFormatter,DateFormatter]");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "title[LowerCaseChanger,IdentityFormatter,DateFormatter]");
 
         assertEquals(3, actions.getConfiguredActions().size());
 
@@ -98,7 +102,7 @@ public class SaveActionsTest {
     @Test
     public void checkMultipleSaveActions() throws IOException {
 
-        SaveActions actions = new SaveActions(true, "pages[PageNumbersFormatter]title[LowerCaseChanger]");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "pages[PageNumbersFormatter]title[LowerCaseChanger]");
 
         List<FieldFormatterCleanup> formatterCleanups = actions.getConfiguredActions();
 
@@ -121,7 +125,7 @@ public class SaveActionsTest {
     @Test
     public void checkMultipleSaveActionsWithMultipleFormatters() throws IOException {
 
-        SaveActions actions = new SaveActions(true, "pages[PageNumbersFormatter,DateFormatter]title[LowerCaseChanger]");
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "pages[PageNumbersFormatter,DateFormatter]title[LowerCaseChanger]");
         List<FieldFormatterCleanup> formatterCleanups = actions.getConfiguredActions();
 
         assertEquals(3, formatterCleanups.size());
@@ -142,4 +146,20 @@ public class SaveActionsTest {
         assertEquals("1--7", entry.getField("pages"));
     }
 
+    @Test
+    public void eraseFormatterRemovesField() {
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, "mont[EraseFormatter]");
+        actions.applySaveActions(entry);
+
+        assertEquals(Optional.empty(), entry.getFieldOptional("mont"));
+    }
+
+    @Test
+    public void getDescriptionCallsFormatter() {
+        Formatter formatter = mock(Formatter.class);
+        FieldFormatterCleanup cleanup = new FieldFormatterCleanup("testField", formatter);
+        when(formatter.getDescription()).thenReturn("Format field %s");
+
+        assertEquals("Format field testField", cleanup.getDescription());
+    }
 }